Finalização do controller e classes de objetos dto

+++ api-personagens/src/main/java/br/ce/anderson/apipersonagens/services/PersonagemService.java
@@ -2,12 +2,16 @@ package br.ce.anderson.apipersonagens.services;
 
 import java.util.List;
 import java.util.Optional;
+import java.util.stream.Collectors;
 
+import org.modelmapper.ModelMapper;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import br.ce.anderson.apipersonagens.model.Personagem;
+import br.ce.anderson.apipersonagens.model.exception.ResourceNotFoundException;
 import br.ce.anderson.apipersonagens.repository.PersonagemRepository;
+import br.ce.anderson.apipersonagens.shared.PersonagemDTO;
 
 @Service
 public class PersonagemService {
@@ -15,28 +19,46 @@ public class PersonagemService {
     @Autowired
     private PersonagemRepository personagemRepository;
 
-    public List<Personagem> obterTodos(){
-        return personagemRepository.findAll();
+    public List<PersonagemDTO> obterTodos(){
+        List<Personagem> personagens = personagemRepository.findAll();
+        return personagens.stream()
+        .map(personagem -> new ModelMapper().map(personagem, PersonagemDTO.class))
+        .collect(Collectors.toList());
     }
 
-    public Optional<Personagem> obterPorId(Integer id){
-        return personagemRepository.findById(id);
+    public Optional<PersonagemDTO> obterPorId(Integer id){
+        Optional<Personagem> p = personagemRepository.findById(id);
+        if(p.isEmpty()){
+            throw new ResourceNotFoundException("Personagem não encontrado.");
+        }
+        PersonagemDTO personagem = new ModelMapper().map(p.get(), PersonagemDTO.class);
+        return Optional.of(personagem);
+
+        //throw new ResourceNotFoundException("Personagem não encontrado");
     }
 
-    public Personagem cadastrar(Personagem p){
-        return personagemRepository.save(p);
+    public PersonagemDTO cadastrar(PersonagemDTO personagemDTO){
+        Personagem personagem = new ModelMapper().map(personagemDTO, Personagem.class);
+        Personagem personagemRetorno = personagemRepository.save(personagem);
+        PersonagemDTO personagemDto = new ModelMapper().map(personagemRetorno, PersonagemDTO.class);
+        return personagemDto;
     }
 
     public void deletar(Integer id){
+        Optional<Personagem> p = personagemRepository.findById(id);
+        if(p.isEmpty()){
+           throw new ResourceNotFoundException("Personagem com id = " + id + "não encontrado.");
+        }
         personagemRepository.deleteById(id);
     }
 
-    public Personagem atualizar(Integer id, Personagem p){
+    public PersonagemDTO atualizar(Integer id, PersonagemDTO p){
 
         //Fazer a validação do campo id aqui
         p.setId(id);
-        personagemRepository.deleteById(id);
-        return personagemRepository.saveAndFlush(p);
+        deletar(id);
+        Personagem personagem = new ModelMapper().map(p, Personagem.class);
+        return new ModelMapper().map(personagemRepository.save(personagem), PersonagemDTO.class); 
     }
